# src/main/resources/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Petstore
  license:
    name: MIT
servers:
  - url: http://localhost:8080/


paths:
  /student/{studentId}:
    get:
      description: This is description
      summary: Fetch student by Id
      operationId: getStudent

      parameters:
        - name: studentId
          in: path
          required: true
          description: The id of the pet to retrieve
          schema:
            type: integer

        - name: studentName
          in: query
          required: true
          description: Student Name
          schema:
            type: string
      

      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      description: This is description
      summary: Fetch student by Id
      operationId: saveStudent

      parameters:
        - name: studentId
          in: path
          required: true
          description: The id of the pet to retrieve
          schema:
            type: integer

      requestBody:
       required: true
       content:
        application/json:
         schema:
          $ref: "#/components/schemas/SaveStudent"

      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
   




components:
  schemas:
    Student:
      type: object
      required:
        - stduentId
        - studentName
      properties:
        stduentId:
          type: integer
          format: int64
        studentName:
          type: string
          nullable: false
        
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string

    SaveStudent:
     type: object
     required:
       - studentId
       - studentName
     properties:
       studentId:
         type: integer
         format: int64
       studentName:
         type: string
         nullable: false